requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gdown==5.2.0
grpcio==1.68.1
hydra-core==1.3.0
huggingface-hub==0.29.2
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
open-metric-learning==3.1.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.1.2
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
soupsieve==2.6
sympy==1.13.3
timm==1.0.15
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.12.12
torch==2.1.1
torchmetrics==1.6.0
torchvision==0.16.1
tqdm==4.67.1
transformers==4.40.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
